refactor: Move state views file
