+++ FuzeChallenge/FuzeChallenge/Modules/MatchesList/MatchListView.swift
@@ -1,6 +1,7 @@
 //
 
 import SwiftUI
+import Combine
 
 struct MatchView: View {
 	@EnvironmentObject private var themeManager: ThemeManager
@@ -141,58 +142,6 @@ struct MatchStatusView: View {
 					topTrailingRadius: 16
 				)
 			)
-	}
-}
-
-
-struct LoadingView: View {
-	@EnvironmentObject private var themeManager: ThemeManager
-
-	var body: some View {
-		ZStack {
-			themeManager.currentTheme.background
-				.ignoresSafeArea()
-			ProgressView()
-				.scaleEffect(2)
-				.tint(themeManager.currentTheme.textPrimary)
-		}
-	}
-}
-
-struct ErrorView: View {
-	private let message: String
-	private let retryAction: () -> Void
-
-	@EnvironmentObject private var themeManager: ThemeManager
-
-	init(message: String, action: @escaping () -> Void) {
-		self.message = message
-		self.retryAction = action
-	}
-
-	var body: some View {
-		ZStack {
-			themeManager.currentTheme.background
-				.ignoresSafeArea()
-
-			VStack {
-				Spacer()
-				Text(message)
-					.font(themeManager.currentTheme.fontBig)
-					.foregroundStyle(themeManager.currentTheme.textSecondary)
-					.lineLimit(2)
-					.padding(24)
-
-				Button("Tentar novamente") {
-					retryAction()
-				}
-				.font(themeManager.currentTheme.fontBig)
-				.foregroundStyle(themeManager.currentTheme.textPrimary)
-				Spacer()
-			}
-			.background(themeManager.currentTheme.background).ignoresSafeArea()
-			.padding()
-		}
 	}
 }
